fix sglang openai return routed experts
